# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prediction Game API",
    "description": "API for wallet-authenticated prediction gameplay, leaderboards, rounds, and predictions. Use Swagger UI to explore endpoints and test requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste a JWT like: Bearer <token>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseErrorResponse"
          },
          {
            "type": "object",
            "properties": {
              "details": {
                "type": "array",
                "description": "Field-level validation details (present on validation errors only)",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "example": "walletAddress"
                    },
                    "message": {
                      "type": "string",
                      "example": "walletAddress is required"
                    }
                  },
                  "required": [
                    "field",
                    "message"
                  ]
                }
              }
            }
          }
        ]
      },
      "BaseErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all API endpoints on failure.",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error class name (e.g. ValidationError, AuthenticationError, NotFoundError)",
            "example": "ValidationError"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error",
            "example": "walletAddress is required"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code for programmatic handling",
            "example": "VALIDATION_ERROR"
          }
        },
        "required": [
          "error",
          "message",
          "code"
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "summary": "List tournaments",
        "description": "Supports optional mode and status filters with offset pagination.",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "schema": {
              "type": "string",
              "enum": [
                "UP_DOWN",
                "LEGENDS"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "UPCOMING",
                "ACTIVE",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated tournament list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tournament"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pagination": {
                          "type": "object",
                          "required": [
                            "limit",
                            "offset",
                            "total"
                          ],
                          "properties": {
                            "limit": {
                              "type": "integer"
                            },
                            "offset": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid mode or status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tournaments"
        ],
        "summary": "Create a tournament (start of the saga lifecycle)",
        "description": "Starts the tournament saga at UPCOMING. Subsequent lifecycle steps\n(join -> lock -> settle -> payout) are validated exclusively in the\nservice layer, not in this route, so out-of-order requests fail with a\nstructured bad-state rejection. Requires authentication.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "mode",
                  "entryFee",
                  "prizePool",
                  "maxParticipants",
                  "startTime",
                  "endTime",
                  "rounds"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "UP_DOWN",
                      "LEGENDS"
                    ]
                  },
                  "entryFee": {
                    "type": "string"
                  },
                  "prizePool": {
                    "type": "string"
                  },
                  "maxParticipants": {
                    "type": "integer"
                  },
                  "startTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "rounds": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tournament created"
          },
          "400": {
            "description": "Invalid tournament parameters"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Lifecycle violation"
          }
        }
      }
    },
    "/api/tournaments/{id}/lock": {
      "post": {
        "tags": [
          "tournaments"
        ],
        "summary": "Lock a tournament (close the join window)",
        "description": "Transitions the tournament UPCOMING -> ACTIVE, freezing registration\nbefore rounds begin. Validated by the saga state machine in the service.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament locked"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Tournament not found"
          },
          "409": {
            "description": "Lifecycle violation (e.g. locking a COMPLETED tournament)"
          }
        }
      }
    },
    "/api/tournaments/{id}/settle": {
      "post": {
        "tags": [
          "tournaments"
        ],
        "summary": "Settle a tournament and pay out winners",
        "description": "Transitions the tournament ACTIVE -> COMPLETED and distributes the prize\npool deterministically based on the tied round leaderboard. The settle +\npayout is atomic and validated by the saga state machine.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Settlement completed with winner allocations"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Tournament not found"
          },
          "409": {
            "description": "Lifecycle violation (e.g. settling an un-locked tournament)"
          }
        }
      }
    },
    "/api/tournaments/{id}/cancel": {
      "post": {
        "tags": [
          "tournaments"
        ],
        "summary": "Cancel a tournament",
        "description": "Transitions the tournament UPCOMING/ACTIVE -> CANCELLED. Terminal\nsuppression: a COMPLETED tournament cannot be cancelled. Validated by\nthe saga state machine in the service layer.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament cancelled"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Tournament not found"
          },
          "409": {
            "description": "Lifecycle violation (e.g. cancelling a COMPLETED tournament)"
          }
        }
      }
    },
...
  }
}